// src/app/models/QualifyListRes.model.ts
export interface QualifyListRes{
    
    id: number;
    idQualify: number;
    idClient:  number;        
    numQualified: number;
    nomConceptDown: string;
    idRelation: number;    
    idApplication: number;
    idCompany: number; 
    idConceptRel: number;
    nomConceptRel: string;
    idConcept: number;
    nomConcept: string;
    indEstatus:string;
    indValue: number;
    indNivel: number;
    

}