// src/app/models/QualifyListMainRes.model.ts
export interface QualifyListMainRes{
    
    id: number; 
    idConceptRisk: number;
    nomCatalogRisk: string;
    maxRisk: number;
    factorx: number;
    sumRisk: number;
    catCalif: number;
    indNivel: number;
    idRelation: number;
    idApplication: number;
    idCompany: number;
    idConceptRel: number;
    idConcept: number;
    indEstatus: string;
    indValue: number;

}